Add /banip and /unbanip

+++ src/main/java/me/oddlyoko/terminator/database/controller/KickManager.java
@@ -90,7 +90,7 @@ public class KickManager {
 			List<KickModel> kicks = new ArrayList<>();
 			while (rs.next()) {
 				// Exist
-				int sanctionId = rs.getInt("sanction_id");
+				long sanctionId = rs.getLong("sanction_id");
 				UUID punishedUuid = UUID.fromString(rs.getString("punished_uuid"));
 				String strPunisherUuid = rs.getString("punisher_uuid");
 				UUID punisherUuid = strPunisherUuid == null ? null : UUID.fromString(strPunisherUuid);
@@ -136,7 +136,7 @@ public class KickManager {
 			List<KickModel> kicks = new ArrayList<>();
 			while (rs.next()) {
 				// Exist
-				int sanctionId = rs.getInt("sanction_id");
+				long sanctionId = rs.getLong("sanction_id");
 				UUID punishedUuid = UUID.fromString(rs.getString("punished_uuid"));
 				String strPunisherUuid = rs.getString("punisher_uuid");
 				UUID punisherUuid = strPunisherUuid == null ? null : UUID.fromString(strPunisherUuid);
